[refactor][NFC] Refactor entire codebase to use new descriptor class (#191)
The new classes have the large advantage of being strongly typed and
giving us a lot of correctness by construction. There is also no longer
a need for an explicit parsing step and it is possible to nicely compose
constructors from parameters, forming the exact descriptor without the
need for string manipulation.

Depends on https://github.com/JLLVM/JLLVM/pull/190