[codegen] Fix operand stack type mismatch after exception handling code
The code in the `OperandStack` incorrectly set the LLVM type of the bottom most operand as well when calling `setHandlerStack`. The method was only used during generation of the exception handling chain which shouldn't affect the operand stack at all as compilation continues afterward as usual.

This PR fixes that issue by removing the `setHandlerStack` method, as it in my opinion is oddly named, and not storing an LLVM type to the operand stack.