[codegen][NFC] Refactor ABI sensitive LLVM attribute application into dedicated functions
Code previously had a lot of code duplication which made applying changes to how ABI attributes are applied a lot more difficult as it required searching all over the codebase.
This PR therefore creates new functions that are called on any call instruction and function to apply attributes important for any methods specific to implementing Java to the call or function.

Additionally, it documents and works around a bug that has in the past caused us many ABI issues with passing integers around. During investigations, I found out that the `gc.statepoint` intrinsic, which calls are converted to, to implement parseable stacks for the GC, are currently not properly created with argument attributes, erasing them all instead. A PR was posted upstream to fix it but won't be part of LLVM until at least LLVM 18.