[vm][codegen] Initial decoupling class object stub generation from JVM to LLVM compilation (#199)
This PR is the first step in removing the `LazyClassLoaderHelper` from
the compilation pipeline entirely. This has the large advantage of
decoupling the compiler from any VM state, making it usable and testable
as a pure "JVM -> LLVM IR" compiler. The way this is achieved is by
replacing the implementations of the methods in `LazyClassLoaderHelper`
to instead call functions with specific name mangling as specified in
`ClassObjectStubMangling.hpp`. The JIT specially recognizes the mangling
scheme and provides the appropriate definitions on-demand via the
`ClassObjectStubDefinitionsGenerator` class.

`LazyClassLoaderHelper` also performed the special task of inlining
these stub definitions if the class object is already loaded. To replace
this functionality while decoupling the system, a
`ClassObjectStubImportPass` LLVM pass was written that imports these
definitions iff the class object is already loaded. LLVMs inliner is
then capable of inlining them in the same fashion as the
`LazyClassLoaderHelper` did previously.

This PR specifically does not 100% switch the mechanism yet, but
currently only does so for field accesses to get early feedback on the
approach and to keep the PR smaller.

To summarize, the advantages of this approach are:
* More deterministic compilation output
* More decoupled design as the compiler no longer relies on the VM state
* Clearer cut responsibilities:
  - the compiler only compiles to LLVM IR
  - the definition generator only provides the stub definitions
- the optimization pass only implements the inlining/importing
optimization

Disadvantages:
* bit more code (but cleaner I'd argue)
* special semantics assigned to specific function names

Depends on https://github.com/JLLVM/JLLVM/pull/198