[NFC] Move `StringInterner` to object library (#204)
Motivation for this change is that we'd like to add a shared library
builder to enforce proper layering and modularization. The
`StringInterner` class has so far been part of the VM module despite not
depending on any part of VM and despite being needed by parts such as
materialization. This creates a dependency of materialization library on
the vm library which is undesirable.

This PR therefore moves it to the object library. The choice where to
move it was somewhat hard, but I chose the object library because 1)
there is precedent here as the class loader is also a kind of "vm
component" that lives there and 2) it only depends on parts of object