[compiler] Implement compilation of OSR method
On-Stack-Replacement (OSR) methods are special versions of a method which start at a specific byte offset that possibly isn't 0. The use-case for such a method is in tiered compilation:
Given an execution of a method (either JIT or Interpreter), it may wish to continue execution within the JIT. An OSR method for a requested offset where execution should continue can then be compiled used.

To facilitate changing the frames of one execution to the new JIT compiled frame, the abstract machine state has to be transferred as well. In the case of the JVM, these are the local variables, the operand stack and the program counter. The latter is already backed into the compilation while the operand and local variables can be passed as an array of `int64_t`s as parameters respectively. The exact calling convention and parameters currently implemented are preliminary and possibly still subject to change.

Since an OSR method may start at any instruction within the Bytecode, compilation had to be restructured a bit. The compilation now does what is essentially a DFS walk that prefers straight-line code, leading to dead code elimination through reachability. This is important for OSR as large parts of the method may not be reachable and do not need to be compiled.