[codegen] Move remaining method calls to the new stub generator
Instance method resolution were the last remaining functionality in the `LazyClassLoaderHelper` class. These were now moved to the stub generator functionality and the class therefore completely removed.
A slight difference in implementation to the `LazyClassLoaderHelper` is that `invokespecial` was split from the other method resolutions as it also requires the class object of the class file containing it for its method resolution. The callers class method was therefore encoded into the mangling scheme.