[CI] Add shared builds builder (#205)
A shared library build has the advantage that they are linked and that
one is able to tell the linker that undefined references are not
allowed. Unlike the static build this ensures that we have clean modules
that do not depend on any other modules, forcing us to have a cleaner
non-monolithic design with less coupling.

Fixes https://github.com/JLLVM/JLLVM/issues/203