[unwind] Import and use `libunwind` as unwinder everywhere (#206)
Many of our most recent issues with unwinding have been caused by bugs
in `libgcc`s implementation of the unwinding directives. More
specifically, it is incapable of retrieving the stack pointer in all
scenarios. This has most recently been exposed by a patch adding and
inlining the stubs for calling static methods:
https://github.com/JLLVM/JLLVM/actions/runs/6461724091/job/17541986427 A
workaround that we used until now is to use the Canonical Frame Address
(CFA), but this is sadly not equal to the stack pointer in every
scenario, causing the crash above.

This PR fixes this and is hopefully a step to easily fixing such issues
in the future by importing and using LLVMs `libunwind` implementation at
ALL times. LLVMs libunwind implementation supports every platform that
we could possibly want to support, is written in proper C++ and
implements unwinding properly as far as I can tell. Only relying on one
unwinder implementation should hopefully make maintenance easier as well

`libunwind` is vendored it in the `3rdParty` library rather than
downloaded via CPM for the following reasons:
* `libunwind` is part of LLVMs Monorepo and cloning it would require
cloning a 2GB repo.
* I would like the capability of modifying `libunwind` to implement
either features or optimizations we want. The copy can therefore be seen
as a fork.
* `libunwind` is very rarely updated upstream for any reason other than
adding new platform support.

One deficiency of LLVMs `libunwind` vs `libgcc` is that it is incapable
of setting callee-saved registers while unwinding. This makes it
impossible for us to put GC pointers in callee-saved registers,
requiring them to be spilled to the stack instead. This has previously
been identified on macOS which uses `libunwind`. Since we'd now use
`libunwind` everywhere, all platforms will suffer from tiny bit worse
code generation. However, given that this issue is an optimization issue
rather than correctness I'd rather use something working than something
fast :) Since we are effectively forking `libunwind` we can add this
capability in the future ourselves.