[materialization] Improve codegen of lambda materialization
The current lambda materialization codegen always prepends the lamda object to the arguments, making it the first argument passed to the trampoline. This is problematic as this requires all caller arguments to be moved to new argument registers or stack slots.

This PR improves the codegen by appending the lambda object to the arguments instead, making the lambda trampoline assembly be a simple move into a register or stack slot.