[materialization] Improve codegen of lambda materialization (#213)
The current lambda materialization codegen always prepends the lamda
object to the arguments, making it the first argument passed to the
trampoline. This is problematic as this requires all caller arguments to
be moved to new argument registers or stack slots.

This PR improves the codegen by appending the lambda object to the
arguments instead, making the lambda trampoline assembly be a simple
move into a register or stack slot.

Example assembly before:

![image](https://github.com/JLLVM/JLLVM/assets/6625526/73b5bd62-55bd-4197-a290-dd97fa358228)
This is a lambda with 6 arguments (excluding `this`) being used in our
JNI interface.

New assembly after this PR:

![image](https://github.com/JLLVM/JLLVM/assets/6625526/fde72e6d-d1b2-4430-9b81-10018dffbf67)