[compiler] Introduce static compilation tool (#209)
Testing the LLVM IR output of the JMV compiler automatically is
currently impossible. Checking it manually is similarly annoying,
requiring one to place breakpoints and dumping the module. A lot of
future tasks will require checking the LLVM IR right after compilation.
Either to check 1) correctness on a much finer grained level than the
execution tests could or 2) to also check the quality of the generated
LLVM IR. A typical example would be checking the generated code for an
intrinsic. A test for `iadd` was added as a showcase how to check LLVM
IR output with `FileCheck`.

A side effect of making compilation a static tool, the compiler has now
become an independent module from the rest of the VM and JIT leading to
cleaner code and concern separation and better maintainability.