[vm] Implement On-Stack-Replacement on `throw`
This is the very first mechanism making use of OSR. This PR completely replaces the manual checks of `activeException` entirely with a system more closely matching how exception handling works in languages like C++ with the addition of JIT and OSR.

The new exception handling now works through the use of "deoptimization" operands in LLVM. These are operands that are put into the stackmap and can be read by the unwinder, similarily to how GC pointers are currently read.
When an exception is thrown, we unwind the stack and search for the frame of a Java method that contains an exception handler for the thrown exception.
If found, the deoptimization operands are used to read the local variables in that frame and transfer them to a new JIT compiled OSR frame that starts at the exception handling.

The large advantage of this mechanism as opposed to the previous is that there is no longer an explicit need to check whether an exception has been thrown after a call. Futhermore, the code for exception handlers is not generated and compiled at all unless there is an `athrow` in a method making it reachable. Since exceptions are naturally "exceptional" cold code it should lead to both better compile time (and therefore VM runtime) + better code generation on the non-exceptional code path.
The last point is made a bit mood by the extra code generation required to make the deopt operands readable but can potentially be improved in the future through liveness analysis checking which local variables are actually needed.