[unwind] Port unwinder code to our fork (#217)
This uses the low level API of `libunwind` rather than the Itanium ABIs
as it is generally more powerful. The unwinding API gained one new
function called `callerFrame` which can be used to get the caller frame.
It is also used to implement `unwindStack` which remains the only
function that can create frames.