[compiler] Add optimization for static final fields (#220)
This PR adds an optimization for for accessing static final fields of
non-reference types.
When generating code for every such field, if the class object is
already initialized, a static constant within the LLVM IR is created
with the value of the field. In all further accesses that value will be
used removing the need to load from memory.

Fixes: #10