[vm][compiler] Introduce globals for class objects and methods
Class objects and pointers to globals have so far been embedded in IR with their address. This is undesirable for many reasons, some described in https://github.com/JLLVM/JLLVM/issues/196.
This PR replaces the pointer embedding with proper external global variables which the JIT linker later resolves to the actual class object and method.