[vm] Simplify JIT layering and dependencies (#224)
Previously it was rather opaque how JIT dylibs were arranged, how, where
and when stubs, compile callbacks and so on are being used. This PR aims
to simplify and document the dependencies within the JIT better by 1)
revamping the important JIT dylibs better names, documenting their
purpose and link order, and 2) using a central stub manager which
dispatches to either the JNI bridge implementation, JITting Java methods
and in the future even able to switch between interpreter and JIT
implementations. `ByteCodeOnDemandLayer` was made redundant in the
process and `JNIImplementationLayer` no longer does lazy compilation as
that is handled externally instead.

Depends on https://github.com/JLLVM/JLLVM/pull/223