[gc] add support for relocating derived pointers
Derived pointers are pointers which are potentially offset from a pointer to an object, the base pointer, by some amount. After relocation, the offset must be preserved by relocating the derived pointer through applying the offset to the relocated base pointer.

This patch implements the required support for relocating such derived pointers within our GC. LLVMs stackmap supplies us with a pair of `(basePtr, derivedPtr)`s which can be used to calculate the offset for the relocation.
A new `WriteableFrameValue` class was introduced on-top of the `FrameValue` class to support writing to frame values making the GC code simpler. Last, I had to also remove the `ObjectRepr` class as I didn't want to expose it as a public class but also wanted `WriteableFrameValue` to be strongly typed. I believe its not too great of a loss and that it violated C++s strict aliasing rules as well.

Lastly, I sadly do not have a regression in the PR as we are currently missing infrastructure to do so. See https://github.com/JLLVM/JLLVM/issues/231
I encountered this missing feature while working on a PR to reduce the amount of deoptimization state in IR.