[compiler] Precisely track types of local variables
The precise type of local variables is required to properly read them out when creating deoptimization state on calls. So far, `ptr addrspace(1)` has been used for this purpose, but this is problematic and has actually led to bugs as LLVM later thought that these were GC pointers.
Not using `ptr addrspace(1)` is also incorrect as actual GC pointers still need to be tracked and kept alive by the deopt.

This PR therefore extends the `ByteCodeTypeChecker` to also calculate the local variables at the start of all basic blocks and adds a `LocalVariables` class, that similar to `OperandStack`, tracks the current type of a local variable throughout compilation. These are also used to generate appropriate loads and stores.