[CI] upgrade to GCC 13
Since we're using C++20 it is more important to use newer compilers than usual. In particular, we've had to deal with bugs in various C++20 features such as https://godbolt.org/z/PzhMTYG3b which are difficult to workaround and very very VERY difficult to debug.
This PR therefore switches our GCC version to GCC 13 which is the newest shipped on GitHub actions.