[unwinder] Workaround Clang 17 build regression (#227)
Due to what looks like a bug and regression in Clang 17 it is currently
not possible to compile JLLVM with Clang 17 or newer. After upgrading my
compiler today I was immediately hit with build failures.

This PR worksaround the Clang bug and leaves a link to the issue tracker
to 1) track the issue and 2) remove it one day if we stop supporting
Clang 17