[compiler] add missing deoptimnization state to builtin exceptions
Deoptimization states should be added to every call within a java method from which the stack may be unwound. Since these methods may throw an exception (e.g. out of memory) or generate a stacktrace, deoptimization state should be added to at the very least be able to retrieve the corresponding JVM bytecode.