[CI] Cache LLVM build immediately upon completion (#237)
The default `actions/cache` action does not save the cache if the
workflow as a whole failed. This is annoying in our case as test
failures in JLLVM should not affect cacheability of the LLVM build.
Using explicit save and restore actions fixes this.

Fixes https://github.com/JLLVM/JLLVM/issues/194