[compiler] Refactor Java method metadata (#239)
Metadata associated with Java code has gotten scattered over time: We
currently have the `JavaMethodMetadata` for class objects and methods,
`DeoptEntry`s for reading local variables and the bytecode offset.

This PR consolidates all metadata into one datastructure again and
creates the following distinction:
* `JavaMethodMetadata` is now used for all metadata that is relevant for
the whole method. It additionally contains extra metadata that can be
specialized for JITted methods, native methods and in the future
interpreter methods.
* Per-PC metadata such as the Java bytecode offset, or the `FrameValue`s
for local variables at a given callsite. This mapping is part of a
mapping contained within the JIT data of `JavaMethodMetadata`.

All low-level metadata access one would ever require should therefore
now be possible with just the `JavaMethodMetadata` (with the exception
of GC pointers which are left as a future task).