[interpreter] add basic integer stack and variable instructions
This PR implements the basic integer operations for manipulating the operand stack and local variables and `iadd`. This makes it possible to first execute some integer instructions such as the `iadd` test case mostly from within the interpreter.

Most importantly, the patch required adding methods to `InterpreterContext` for manipulating the operand stack and local variables.