[gc] add new interface for adding external roots to the GC
There are other heaps such as the `ClassLoader` containing Java objects which may refer to Java objects on the GC. To be able to mark such objects as reachable and update references as they are relocated, a new interface is added in this PR called `RootProvider` which adds convenient methods for adding callbacks for root traversal. As a POC, the class objects in the `ClassLoader` are added as root sets.

Sadly, I failed to write a meaningful test as part of this PR as we are currently incapable of producing scenarios where `ClassObject`s point to Java objects in the GC heap with our current native methods in `Class`.

A later PR will use this functionality to implement GC support with the interpreter.

Fixes https://github.com/JLLVM/JLLVM/issues/197