[interpreter] initial interpreter infrastructure (#241)
This PR implements the first initial infrastructure for implementing a
Bytecode Interpreter. It most notably hooks into the existing JIT system
to generate stubs for compiling interpreter methods using the C calling
convention as is currently the case for the JIT and integrates with
`JavaFrame` to implement reading metadata such as local variables, the
operand stack and the bytecode offset.

The actual interpreter method does not yet implement any JVM
instruction. Rather, it only implements On-Stack-Replacement from the
interpreter to JITted code, and uses this mechanism as a fallback to
continue execution of any method containing bytecode not yet implemented
in the interpreter. This allows an incremental implementation strategy
of the interpreter.

To retain test-coverage, `-Xjit` and `-Xint` developer options were
added that control whether code should only be executed by the JIT or
interpreter (if possible). This way our current execution tests can be
used for both the interpreter and the JIT to test their correctness.
Since the Interpreter escapes to the JIT on unimplemented instructions,
no tests failures currently occur when running with `-Xint`.
Nevertheless, if an instruction was implemented incorrectly by the
interpreter, we would see failing tests.

This PR contains only the bare minimum of code required to start
iterating.
 Things left to do are:
 * actually implementing at least one instruction
 * handling interpreter frames in the GC
* implementing OSR from JIT to the Interpreter (for e.g. exception
handling)