[interpreter] add support for `new`
This PR adds support for the 'new' instruction, making it possible to create Java objects from within the interpreter.
Since Java objects are now in the interpreter context, garbage collection support was added to the interpreter as well. This is implemented using the new `RootProvider` API, making it possible to add the interpreter's operand stack and local variables as roots for marking and relocation.