[vm][compiler] Introduce globals for interned strings (#254)
This PR removes the embedding of pointers to strings interned at compile
time with proper external global variables which the JIT linker later
resolves.
This in term removes the need for `StringInterner` for the generation of
LLVM IR and all associated facets.
Altogether reducing coupling and leading to a cleaner design.