[object] Add missing superinterfaces lookup in interface method resolution
The JVM spec states that as a last step, the method should be resolved in any of the interface methods as long as it is not static or private.
This was previously not done for unknown reasons.

This patch implements the missing lookup step. While touching the code, `getMethod` was redefined to only do a lookup in the current class, as that is the more common operation. The previous `getMethod` was renamed to `getMethodSuper`