[codegen] Correctly deduce local variable types in exception handlers
The bytecode type checker was only propagating types of local variables to exception handler at the start of the exception handler. This would lead to incorrectly deducing a type of a local variable where it should be inaccessible instead.
This PR fixes the issue by propagating local variable types for every operation that may throw an exception.
As a side-effect of this change, `Code` now also builds an interval tree for conveniently and quickly finding all exception handlers for a given bytecode offset.