[materialization] Implement OSR to interpreter frames (#249)
So far it has only been possible to perform OSR into a JITted frame.
This PR also adds the capability to OSR into the interpreter and
switches the exception handler to do so by default.

This is achieved by creating an `InterpreterOSRLayer` which compiles a
version of the interpreter entry stub which initializes the abstract
machine state from an `OSRState` instance.