[interpreter] Implement stack operations (#261)
These operations are the few operations that operate on raw memory
rather than requiring a specific typed operand. Most importantly, these
behave different in the spec depending on whether the values are `long`
or `double` or others. In a previous PR we started using two operand
stack slots for `long` and `double` (which is also supported by the
`Code` attribute), making the operations in "Form 1" as the spec calls
it, identical to when operating on `double` or `long`.