[object] Treat array as proper Object subclass (#257)
Up until now it was impossible to call object methods on arrays via
`invokevirtual` and related as arrays did not subclass Object in the
created datastructures. Similarly, the mangling was incapable of
representing array types.

This PR fixes these issues, making array a proper subclass of `Object`,
including V-tables, and changing the mangling to use descriptor syntax
rather than a class name.

---------

Co-authored-by: nemegyeb <80166690+nemegyeb@users.noreply.github.com>