[debuginfo] Extract and consistently use `TrivialDebugInfoBuilder` (#262)
Various methods for generating stubs with LLVM IR where either missing
debug info or repeating lots of code to construct the minimum amount of
debug info required.

This PR extracts the `TrivialDebugInfoBuilder` out of code generation
and makes it available as class and library for reuse.

Looking at the callstack in the debugger is also improved by this PR as
methods that previously didn't emit debug info now do.

![Screenshot from 2023-12-09
12-06-06](https://github.com/JLLVM/JLLVM/assets/6625526/7c2976b5-bc8f-46a6-9ee2-f0e955bebda8)