[interpreter] implement arithmetic operations (#263)
This PR reuses the `MultiTypeImpls` class to implement the basic
arithmetic operations generically on all types supported by the JVM.
Exceptions to this rule are division and remainder, which have differing
semantics in the case of `rhs == 0` among other things.

I opted not to implement throwing an `ArithmeticException` yet as it is
not done in the JIT yet either. Implementing both at the same time makes
the most sense to me.