[interpreter] Implement field accesses
This PR implements accessing both instance and static fields of Java Objects. The implementation simply finds the given field in the class object and then uses memcpy to load the corresponding value according to the descriptor. The address is either formed via an offset from the object if an instance field or directly returned by the field if static.