[interpreter] Implement almost all array operations
This PR implements all array operations but `multianewarray` and `newarray`. As many of these operations are capable of throwing different exceptions where the JIT compiler already uses nice messages, the methods creating these exception objects and throwing them were moved to their own methods in `VirtualMachine` and reused.
Since the version of them throwing is nicer and easier to use, the JIT code was also adjusted to expect these methods to throw the exception rather than create the exception object.