[interpreter] Implement `invoke*` instructions
This PR largely builds on the runtime refactor to introduce a new interface to all executors, allowing all methods to be called with the so-called interpreter calling convention as well.

The interpreter calling convention allows easily calling the resolved method using a reference to the operand stack. A special adaptor layer has been implemented to translate from the interpreter calling convention to the JIT calling convention, allowing the interpreter to call methods implemented in the JIT as well.

Note that this PR is only the minimal working version and is not very optimized. In particular, the interpreter does not cache any method resolutions or selections, nor does it yet directly implement the interpreter calling convention yet, instead relying on the adaptor.
These are seen as optimizations for future PRs.