[interpreter] Implement constant instructions
This PR implements all instructions that push constants onto the operand stack. LDC variants only implement the info structs that the JIT implements as well.

Note that `fmod` was added in the JIT symbols due to LLVM generating a call to it when compiling an OSR version triggered by this patch.