[interpreter] Optimize interpreter calling convention calls into the interpreter
Our VM currently uses two calling conventions: The JIT calling convention and the Interpreter calling convention.
The latter is used by the interpreter to make calls to other methods.

Prior to this PR, if the interpreter called another method in the interpreter-tier an inefficiency occurred: The interpreter currently only implements the JIT calling convention and made use of adaptors to translate between the interpreter and JIT calling convention. The JIT CC interpreter entry then translated back from the JIT calling convention to the interpreter state.

This PR fixes that by creating a new entry method into the interpreter. It is implemented in LLVM IR compiled only once on startup and is callable with just a method object and an arguments array. This entry is then used for ALL methods when using the interpreter CC, meaning no LLVM IR or adaptors have to ever be generated for interpreter to interpreter calls.

The entry uses the same method metadata and deopt operands as interpreter entries generated so far to remain compatible with the rest of the VM.

The only larger sideeffect of this change is that OSR had to now also be implemented for the interpreter calling convention. This was not needed so far as all the methods being replaced have so far followed the JIT calling convention. The method metadata now additionally contains the calling convention used and was reworked a bit to be more memory efficient, easier to use and work with the new single interpreter entry.