[interpreter] Fix UB in `*neg` implementations
The implementations of `ineg` and `lneg` incorrectly used the signed type when performing the negation. In C++ signed overflow is undefined behavior which is luckily caught by UB sanitizer.

This PR fixes that issue by simply using the unsigned type to perform the negation. Negation on unsigned is specified to behave the same as signed but without the undefined behavior.