[interpreter] fix bug in `put*` of wide types (#275)
The implementations of field access methods made use of `pushRaw` and
`popRaw` as the types being operated on are runtime dependent on the
type of the field being accessed. However, using these methods is
dangerous, as this requires manually handling wide types such as `long`
and `double`. This has happened in the implementations of `putstatic`
and `putfield` where the order of `popRaw` calls was incorrect, leading
to always storing 0 in the field.

This PR fixes that issue by adding new overloads for `push` and `pop`
which take a field descriptor as parameter. This adds a safe and
convenient version of `pop` and `raw` while still being capable of
specifying types at runtime.