[ci][cmake] Workaround "bug" in LLD (#286)
The combination of exception handling, optimizations creating empty
functions and using LLD exposes an upstream issue in the toolchain which
leads to `libgcc` and `libunwind` crashing on malformed data:
https://github.com/llvm/llvm-project/issues/61434

This has also lead to our CI crashing:
https://github.com/JLLVM/JLLVM/actions/runs/7346263904/job/20000726099

The issue is most easily worked around by not using LLD in release
builds. Other fixes such as reducing our use of `llvm_unreachable` or
checking whether there are compiler options to avoid emitting zero-sized
functions can be considered in the future.

Until this issue is fixed upstream, change our CI to not use LLD in
fully optimized builds and add code to cmake to warn when using release
builds with LLD to avoid the issue.