[interpreter] Implement constant instructions (#278)
This PR implements all instructions that push constants onto the operand
stack. LDC variants only implement the info structs that the JIT
implements as well.

Note that `fmod` was added in the JIT symbols due to LLVM generating a
call to it when compiling an OSR version triggered by this patch.

Depends on https://github.com/JLLVM/JLLVM/pull/275 and
https://github.com/JLLVM/JLLVM/pull/286