[native] Implement `clinit` of `FileInputStream`
This is required by the startup sequence of `System` as it calls `new` on `FileInputStream`.

This has so far not been an issue due to a bug in the JIT compiler where `new` does not trigger class object initialization.
The interpreter does implement this correctly and therefore requires class initialization of `FileInputStream` to succeed.