[interpreter] Undo loaded class optimization in `checkcast` and `instanceof` (#292)
The optimization is sadly somewhat problematic in `checkcast`
specifically as it makes it impossible/hard to create a proper error
message for the exception in this case.

Furthermore, the spec arguably requires loading of the class and
throwing any class loading related exceptions if it were to fail. This
PR therefore undoes the optimization for the time being.