[vm][NFC] Refactor JIT and Interpreter classes (#266)
The JIT class has become somewhat too overloaded, taking on not only the
responsibility of executing Java Bytecode through compiling to LLVM but
taking on essentially every task related to compiling LLVM IR.

This quickly leads to confusing interdependencies where e.g. the
interpreter needing to compile small IR stubs has a dependency on the
JIT or the JIT being responsible for performing OSR.

This PR fixes this state by refactoring our classes and their
dependencies. A 'Runtime' class was extracted which abstracts over the
specific 'Executor' (JIT, Interpreter or JNI) that is being used to
execute a method. The 'Executor' and 'OSRTarget' interfaces are used for
generically interacting with the various tiers of execution.