[class] Optimize `*switch` instruction representations (#301)
`LookupSwitch` and `TableSwitch` were the only two non-trivial types in
the bytecode leading to the variant as a whole to be non-trivial.
Furthermore, representing TableSwitch as a LookupSwitch made the former
`O(log n)` in the interpreter rather than `O(1)` now.

This PR fixes these issues by getting rid of the `std::vector` used in
both instructions and replacing it with a `BigEndianArrayRef`
implemented as part of this PR that refers to the storage in the `Code`
object for the tables.

While I failed to measure any performance benefit in startup, the
executable size of a release build with debug info did reduce by 2 MB. I
suspect this is partly due to the removal of a few exception handlers in
the interpreter loop