[support] Fix UB in BitArrayRef
The code triggers UB in the case that the index used is 32 or larger. The reason is the use of the `1` constant which defaults to `int` in C++.
Shifting that constant by more than 31 bits is UB and triggered an assertion when built with UBSAN.

This PR fixes that by creating a nicely named constant which is the value 1 cast to the integer type.