[object] Refactor `GCRootRef`
Key motivations for this PR are a few deficiencies when using 'GCRootRef':
* It is not default-constructible or constructible from a nullptr. This is sadly required to make `GCRootRef` equal to `jobject` in the JNI.
* Using `GCRootRef<Array>` is hard to use as it's cumbersome to access the array elements.
* `GCRootRef` is implicitly convertible to `T*` which has some nasty effects during operator overloading. More concretely, it makes e.g. `operator[]` impossible to implement due to ambiguity.

This PR addresses all three of these issues.

For the first, `GCRootRef`s have a new state where it does not have a root. This is similar to the state of a moved from `GCUniqueRoot` previously. From a user-perspective a rootless `GCRootRef` is equivalent to a root containing a nullptr with the only exception of object reassignment.
Assigning a `T*` to a `GCRootRef` is only possible if it contains a root. The `operator=` that previously did object reassignment has therefore been changed to an explicit method to make this clearer.

For array types a new `GCRootRefTrait` was introduced which injects extra `begin`, `end` and `operator[]` methods when `T` is an array. Through the use of a proxy they can be safely used regardless of any GCs inbetween.

Regarding implicit pointer conversions: The implicit pointer conversion was originally introduced as a matter of ergonomics and reuse. It was simply cumbersome to have to decide whether a method should take a `T*` or `GCRootRef` and it was decided that the former is a more canonical former that the latter converts to.
Given various nasty issues with pointer types in C++, I believe this to nowadays be more problematic than useful.
To nevertheless address the parameter type issue, adaptor classes were introduced which can be implicitly constructed from both pointers and `GCRootRef`s. APIs commonly used with kinds can use these as parameters to make both kinds of calls work.
In a sense these adaptors can be thought of as a "opt-in implicit pointer conversion" for `GCRootRef` as opposed to the "always implicit pointer conversion" from before.