[native] Implement `isAlive` for some JDK 17 distros (#310)
The reference OpenJDK source has `isAlive` as a native method that has
to be implemented in native code:
https://github.com/openjdk/jdk/blob/dfacda488bfbe2e11e8d607a6d08527710286982/src/java.base/share/classes/java/lang/Thread.java#L1053

Some distributions of the OpenJDK (e.g. Oracle, Amazon corretto) seem to
ship their JDK 17s with the preview for virtual threads
(https://github.com/openjdk/jdk/commit/9583e3657e43cc1c6f2101a64534564db2a9bd84)
which changed the implementation to be completely in Java.

This PR implements `isAlive` to also support OpenJDKs that do not ship
the preview (e.g. the builds one gets on Ubuntu or Fedora from their
package managers). In the case that `isAlive` is not a `native` method,
the implementation is simply ignored making this PR a noop on such
distributions.