[JNI] Create basic unit testing for JNI (#304)
This PR creates the most basic test infrastructure for implementing the
Java Native Interface. It basically consists of a fixture containing a
`VirtualMachine` instance and a `JNIEnv` that tests interact with. For
this purpose, `VirtualMachine` was made easily bootable from just a
single static `create` method.

Furthermore, a cmake macro was created for compiling Java files to class
files. These will be able to serve as interesting inputs for the JNI
interfaces (e.g. for accessing fields).