[JNI] Implement static field access
This PR implements the basic happy path of the JNI methods used to access static fields.
These consist of `GetStaticFieldID` to fetch a static field and then the various `GetStatic*Field` and `SetStatic*Field` methods used to read or write to static fields.

The most important part of this PR is the infrastructure setup to conveniently implement these functions. Since we need to perform type conversions from JNI to JLLVM types, metaprogramming facilities were implemented that automatically convert between lambdas using JLLVM types to the corresponding JNI types.

The current implementation does not fully implement these methods in the error case as throwing and catching exceptions is not yet implemented for the JNI.