[interpreter] Optimize OSR entry into the interpreter (#299)
Performing OSR into the interpreter currently needlessly always compiles
a new stub with LLVM for every bytecode offset, method and calling
convention. This is needless compilation time and memory overhead that
likely outweigh any benefits of the specialized compiled entry.

This PR therefore replaces that implementation of interpreter OSR
entries with general OSR entries that initialize their state from just
the OSR State passed in regardless of method and bytecode offset. This
is analogous to the recent work on a single interpreter entry.

Since OSR may lead to requiring to replace frames with different calling
conventions, versions of the OSR entry are compiled once on startup for
every possible return type and then selected during OSR.